Update: 銘柄から蔵名の補完をStimulus化した
