--- app/javascript/controllers/complete_meigara_controller.js
@@ -1,7 +1,0 @@
-import { Controller } from "@hotwired/stimulus"
-
-// Connects to data-controller="complete-meigara"
-export default class extends Controller {
-  connect() {
-  }
-}
